+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 ## 2.12.0 - 2026-07-13
-- Upgrade Otel dependencies to [1.43.0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.43.0) / [0.64b0](https://github.com/open-telemetry/opentelemetry-python-contrib/releases/tag/v0.64b0)
+- Upgrade Otel dependencies to [1.44.0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.44.0) / [0.65b0](https://github.com/open-telemetry/opentelemetry-python-contrib/releases/tag/v0.65b0)
 
 ## 2.11.0 - 2026-06-09
 - Upgrade Otel dependencies to [1.42.1](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.42.1) / [0.63b1](https://github.com/open-telemetry/opentelemetry-python-contrib/releases/tag/v0.63b1)